deinterleave

+++ PhotoZ/Source/MainControllerDAP.cpp
@@ -23,7 +23,7 @@ char* i2txt(int);
 void MainController::setResetOnset(const char *txt)
 {
 	
-	float value=atof(txt);
+	double value=atof(txt);
 	//std::cout << txt;
 	if(value<0) value=0;
 
@@ -36,7 +36,7 @@ void MainController::setResetOnset(const char *txt)
 //=============================================================================
 void MainController::setResetDuration(const char *txt)
 {
-	float value=atof(txt);//temporary might need to change back to atoi
+	double value=atof(txt);//temporary might need to change back to atoi
 	if(value<0) value=0;
 
 	dc->reset->setDuration(value);
@@ -48,7 +48,7 @@ void MainController::setResetDuration(const char *txt)
 //=============================================================================
 void MainController::setShutterOnset(const char *txt)
 {
-	float value=atof(txt);
+	double value=atof(txt);
 	if(value<0) value=0;
 
 	dc->shutter->setOnset(value);
@@ -60,7 +60,7 @@ void MainController::setShutterOnset(const char *txt)
 //=============================================================================
 void MainController::setShutterDuration(const char *txt)
 {
-	float value=atof(txt);
+	double value=atof(txt);
 	if(value<0) value=0;
 
 	dc->shutter->setDuration(value);
@@ -72,7 +72,7 @@ void MainController::setShutterDuration(const char *txt)
 //=============================================================================
 void MainController::setStiOnset(int ch, const char *txt)
 {
-	float value=atof(txt);
+	double value=atof(txt);
 	if(value<0) value=0;
 	//value += 3.5*(1/(double)Camera::FREQ[dc->getCameraProgram()]);
 
@@ -94,7 +94,7 @@ void MainController::setStiOnset(int ch, const char *txt)
 //=============================================================================
 void MainController::setStiDuration(int ch, const char *txt)
 {
-	float value=atof(txt);
+	double value=atof(txt);
 	if(value<0) value=0;
 
 	if(ch==1)
@@ -115,7 +115,7 @@ void MainController::setStiDuration(int ch, const char *txt)
 //=============================================================================
 void MainController::setAcquiOnset(const char *txt)
 {
-	float value=atof(txt);
+	double value=atof(txt);
 	if(value<0) value=0;
 
 	dc->setAcquiOnset(value);
